feat: Protect transactions endpoint

+++ backend/app/api/transactions.py
@@ -2,9 +2,10 @@ from fastapi import APIRouter, Depends, HTTPException, status
 from sqlmodel import Session, col, func, select
 from uuid import UUID
 from datetime import datetime
-from app.models import Transaction
+from app.models import Category, Transaction, User
 from app.schemas import TransactionCreate, TransactionResponse, TransactionUpdate
 from app.database import get_db_session
+from app.security import get_current_user
 
 router = APIRouter(tags=["transaction"])
 
@@ -15,9 +16,22 @@ router = APIRouter(tags=["transaction"])
     status_code=status.HTTP_201_CREATED,
 )
 def create_transaction(
-    transaction: TransactionCreate, session: Session = Depends(get_db_session)
+    transaction: TransactionCreate,
+    session: Session = Depends(get_db_session),
+    current_user: User = Depends(get_current_user),
 ):
-    db_transaction = Transaction(**transaction.model_dump())
+    category = session.get(Category, transaction.category_id)
+    if not category:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND, detail="Category not found"
+        )
+
+    if category.user_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Not authorized to user this category",
+        )
+    db_transaction = Transaction(**transaction.model_dump(), user_id=current_user.id)
     session.add(db_transaction)
     session.commit()
     session.refresh(db_transaction)
@@ -26,18 +40,43 @@ def create_transaction(
 
 @router.get("/transactions", response_model=list[TransactionResponse])
 def get_transactions(
-    skip: int = 0, limit: int = 10, session: Session = Depends(get_db_session)
+    skip: int = 0,
+    limit: int = 10,
+    category_id: UUID | None = None,
+    start_date: datetime | None = None,
+    end_date: datetime | None = None,
+    session: Session = Depends(get_db_session),
+    current_user: User = Depends(get_current_user),
 ):
-    transactions = session.exec(select(Transaction).offset(skip).limit(limit)).all()
+    query = select(Transaction).where(Transaction.user_id == current_user.id)
+
+    if category_id:
+        query = query.where(Transaction.category_id == category_id)
+    if start_date:
+        query = query.where(Transaction.date >= start_date)
+    if end_date:
+        query = query.where(Transaction.date <= end_date)
+
+    query = query.order_by(col(Transaction.date).desc())
+    transactions = session.exec(query.offset(skip).limit(limit)).all()
     return transactions
 
 
 @router.get("/transactions/{transaction_id}", response_model=TransactionResponse)
-def get_transaction(transaction_id: UUID, session: Session = Depends(get_db_session)):
+def get_transaction(
+    transaction_id: UUID,
+    session: Session = Depends(get_db_session),
+    current_user: User = Depends(get_current_user),
+):
     transaction = session.get(Transaction, transaction_id)
     if not transaction:
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND, detail="transaction not found"
+        )
+    if transaction.user_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Not authorized to access this transaction",
         )
     return transaction
 
@@ -47,12 +86,35 @@ def update_transaction(
     transaction_id: UUID,
     transaction_update: TransactionUpdate,
     session: Session = Depends(get_db_session),
+    current_user: User = Depends(get_current_user),
 ):
     transaction = session.get(Transaction, transaction_id)
     if not transaction:
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND, detail="Transaction not found"
         )
+
+    if transaction.user_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Not authorized to access this transation",
+        )
+
+    if (
+        transaction_update.category_id
+        and transaction_update.category_id != transaction.category_id
+    ):
+        new_category = session.get(Category, transaction_update.category_id)
+        if not new_category:
+            raise HTTPException(
+                status_code=status.HTTP_404_NOT_FOUND, detail="Category not found"
+            )
+        if new_category.user_id != current_user.id:
+            raise HTTPException(
+                status_code=status.HTTP_403_FORBIDDEN,
+                detail="Not authorized to use this category",
+            )
+
     transaction_data = transaction_update.model_dump(exclude_unset=True)
     for key, value in transaction_data.items():
         setattr(transaction, key, value)
@@ -64,44 +126,36 @@ def update_transaction(
 
 @router.delete("/transactions/{transaction_id}")
 def delete_transaction(
-    transaction_id: UUID, session: Session = Depends(get_db_session)
+    transaction_id: UUID,
+    session: Session = Depends(get_db_session),
+    current_user: User = Depends(get_current_user),
 ):
     transaction = session.get(Transaction, transaction_id)
     if not transaction:
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND, detail="Transaction not found"
         )
+    if transaction.user_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Not authorized to delete this transaction",
+        )
     session.delete(transaction)
     session.commit()
     return {"message": "Transaction deleted successfully"}
 
 
-@router.get("/transactions/user/{user_id}", response_model=list[TransactionResponse])
-def get_transactions_by_user(
+@router.get("/transactions/user/summary")
+def get_my_transaction_summary(
     user_id: UUID,
-    category_id: UUID | None = None,
-    start_date: datetime | None = None,
-    end_date: datetime | None = None,
     session: Session = Depends(get_db_session),
+    current_user: User = Depends(get_current_user),
 ):
-    query = select(Transaction).where(Transaction.user_id == user_id)
-    if category_id:
-        query = query.where(Transaction.category_id == category_id)
-    if start_date:
-        query = query.where(Transaction.date >= start_date)
-    if start_date:
-        query = query.where(col(Transaction.date) <= end_date)
-    transactions = session.exec(query).all()
-    return transactions
-
-
-@router.get("/transactions/user/{user_id}/summary")
-def get_transaction_summary(user_id: UUID, session: Session = Depends(get_db_session)):
 
     total_expenses = (
         session.exec(
             select(func.sum(Transaction.amount)).where(
-                Transaction.user_id == user_id, Transaction.amount > 0
+                Transaction.user_id == current_user.id, Transaction.amount > 0
             )
         ).first()
         or 0
@@ -114,12 +168,26 @@ def get_transaction_summary(user_id: UUID, session: Session = Depends(get_db_ses
 
 @router.get("/transactions/category/{category_id}/summary")
 def get_transsaction_summary_by_category(
-    category_id: UUID, session: Session = Depends(get_db_session)
+    category_id: UUID,
+    session: Session = Depends(get_db_session),
+    current_user: User = Depends(get_current_user),
 ):
+    category = session.get(Category, category_id)
+    if not category:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND, detail="Category not found"
+        )
+    if category.user_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Not authorized to access this category",
+        )
     total_expenses = (
         session.exec(
             select(func.sum(Transaction.amount)).where(
-                Transaction.category_id == category_id, Transaction.amount > 0
+                Transaction.category_id == category_id,
+                Transaction.user_id == current_user.id,
+                Transaction.amount > 0,
             )
         ).first()
         or 0
@@ -131,9 +199,29 @@ def get_transsaction_summary_by_category(
     "/transactions/category/{category_id}", response_model=list[TransactionResponse]
 )
 def get_transactions_by_category(
-    category_id: UUID, session: Session = Depends(get_db_session)
+    category_id: UUID,
+    session: Session = Depends(get_db_session),
+    current_user: User = Depends(get_current_user),
 ):
+    category = session.get(Category, category_id)
+
+    if not category:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail="Category not found",
+        )
+    if category.user_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Not authorized to access this category",
+        )
+
     transactions = session.exec(
-        select(Transaction).where(Transaction.category_id == category_id)
+        select(Transaction)
+        .where(
+            Transaction.category_id == category_id,
+            Transaction.user_id == current_user.id,
+        )
+        .order_by(col(Transaction.date).desc())
     ).all()
     return transactions
